+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Next Release
+
+### New features
+
+- Added `\Wizaplace\SDK\Order\OrganisationOrderService::getOrder`
+
 ## 1.41.0
 
 Compatible with Wizaplace 1.27.0